juliao_api/pyproject.toml
[tool.poetry]
name = "juliao-api"
version = "0.1.0"
description = "Backend API for Julião personal finance app"
authors = ["Your Name <you@example.com>"] # Placeholder, can be updated later
readme = "README.md" # Assuming a README will be added in juliao_api later

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1" # Using a recent version, can be adjusted
uvicorn = {extras = ["standard"], version = "^0.23.2"}
pydantic = "^2.5.2"
pydantic-settings = "^2.1.0"
sqlalchemy = "^2.0.23" # Or sqlmodel if preferred by the team
psycopg2-binary = "^2.9.9" # For PostgreSQL, use asyncpg for async
alembic = "^1.13.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
APScheduler = "^3.10.4"
sqlmodel = "^0.0.14"
httpx = "^0.25.1" # For fetching JWKS

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
black = "^23.11.0"
ruff = "^0.1.6" # Replaces flake8, isort, etc.

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
